refactor: rename to BookingDataService.
